crawlers/requirements.txt
# Scrapy and extensions
scrapy==2.11.0
scrapy-playwright==0.0.33  # For JavaScript sites
scrapy-redis==0.7.3  # For distributed crawling
scrapyd==1.4.3  # For deployment
scrapyd-client==1.2.3  # For deploying to Scrapyd

# Cloud storage
boto3==1.34.13  # AWS S3
google-cloud-storage==2.13.0  # Google Cloud Storage

# Data processing
itemadapter==0.8.0
python-dateutil==2.8.2

# API client
httpx==0.26.0
tenacity==8.2.3  # Retry logic

# Monitoring
sentry-sdk==1.40.0

# Development
ipython==8.20.0
scrapy-fake-useragent==1.4.4
